feat: fix a few bag and refactoring.
