+++ src/main/scala/jp/ed/nnn/movieplayer/Main.scala
@@ -66,15 +66,15 @@ class Main extends Application {
 
     tableView.getColumns.setAll(fileNameColumn, timeColumn, deleteActionColumn)
 
-    val toolBar = ToolbarCreator.create(mediaView, tableView, timeLabel, primaryStage)
-
     val baseBorderPane = new BorderPane()
+    val scene = new Scene(baseBorderPane, mediaViewFitWidth + tableMinWidth, mediaViewFitHeight + toolBarMinHeight)
+    val toolBar = ToolbarCreator.create(mediaView, tableView, timeLabel, scene, primaryStage)
+
     baseBorderPane.setStyle("-fx-background-color: Black")
     baseBorderPane.setCenter(mediaView)
     baseBorderPane.setBottom(toolBar)
     baseBorderPane.setRight(tableView)
 
-    val scene = new Scene(baseBorderPane, mediaViewFitWidth + tableMinWidth, mediaViewFitHeight + toolBarMinHeight)
     scene.setFill(Color.BLACK)
     mediaView.fitWidthProperty().bind(scene.widthProperty().subtract(tableMinWidth))
     mediaView.fitHeightProperty().bind(scene.heightProperty().subtract(toolBarMinHeight))
